Migrate TeamID to int64 and enable CKP server

+++ cookiefarm/server/internal/websockets/event.go
@@ -50,7 +50,7 @@ func FlagHandler(event Event, client *Client) error {
 	logger.Log.Debug().
 		Int("client", client.Number).
 		Str("flag", flag.FlagCode).
-		Uint16("team", flag.TeamID).
+		Int64("team", flag.TeamID).
 		Str("service name", flag.ServiceName).
 		Uint16("port service", flag.PortService).
 		Msg("Flag received and sent to DB")
